+++ docsource/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added 
 
-
+- glossary
+- show_api_keys
 - public folder to hold webpage documents for gitlab.
 - docs folder to hold webpage documents for github.
 - Set up environment page. 
@@ -12,6 +13,7 @@
 
 ### Changed
 
+- get_census --> census
 - Reorganized the imports
 - docs/source --> docsource
 - Moved Makefiles into docsource folder.